Add menu core type
