--- src/golem-ui/src/application/menu/cores.rs
@@ -1,4 +1,0 @@
-// pub mod download;
-//
-// pub mod list;
-// pub use list::*;
