+++ src/golem-ui/src/core_manager.rs
@@ -1,10 +1,11 @@
 use std::path::Path;
+use std::time::SystemTime;
 
 use byteorder::{LittleEndian, ReadBytesExt};
 
 use mister_fpga::config::{Config, HdmiLimitedConfig, VgaMode};
 use mister_fpga::core::file::SdCard;
-use mister_fpga::core::MisterFpgaCore;
+use mister_fpga::core::{MenuCore, MisterFpgaCore};
 use mister_fpga::fpga::user_io::{ButtonSwitches, UserIoButtonSwitch};
 use mister_fpga::fpga::MisterFpga;
 use one_fpga::core::SaveState;
@@ -30,50 +31,23 @@ impl CoreManager {
 
     /// Create a core for the current FPGA configuration.
     fn create_core(&mut self, is_menu: bool) -> Result<GolemCore, String> {
-        let mut core = MisterFpgaCore::new(self.fpga.clone())
-            .map_err(|e| format!("Could not instantiate Core: {e}"))?;
+        let mut core = if is_menu {
+            MenuCore::new(self.fpga.clone())
+                .map(GolemCore::new)
+                .map_err(|e| format!("Could not instantiate Core: {e}"))?
+        } else {
+            MisterFpgaCore::new(self.fpga.clone())
+                .map(GolemCore::new)
+                .map_err(|e| format!("Could not instantiate Core: {e}"))?
+        };
 
         let options = Config::base().into_inner();
 
-        core.init()?;
-        core.init_video(&options, is_menu)?;
-        core.send_volume(0)?;
+        core.init().map_err(|e| e.to_string())?;
+        core.set_volume(0).map_err(|e| e.to_string())?;
+        core.set_rtc(SystemTime::now()).map_err(|e| e.to_string())?;
 
-        let mut switches = UserIoButtonSwitch::new();
-        if options.vga_scaler == Some(true) {
-            switches |= ButtonSwitches::VgaScaler;
-        }
-        if options.vga_sog == Some(true) {
-            switches |= ButtonSwitches::VgaSog;
-        }
-        if options.composite_sync == Some(true) {
-            switches |= ButtonSwitches::CompositeSync;
-        }
-        if options.vga_mode() == VgaMode::Ypbpr {
-            switches |= ButtonSwitches::Ypbpr;
-        }
-        if options.forced_scandoubler() {
-            switches |= ButtonSwitches::ForcedScandoubler;
-        }
-        if options.hdmi_audio_96k() {
-            switches |= ButtonSwitches::Audio96K;
-        }
-        if options.dvi_mode() {
-            switches |= ButtonSwitches::Dvi;
-        }
-        match options.hdmi_limited() {
-            HdmiLimitedConfig::Limited => switches |= ButtonSwitches::HdmiLimited1,
-            HdmiLimitedConfig::LimitedForVgaConverters => switches |= ButtonSwitches::HdmiLimited2,
-            _ => {}
-        }
-        if options.direct_video() {
-            switches |= ButtonSwitches::DirectVideo;
-        }
-
-        core.spi_mut().execute(switches).unwrap();
-        core.send_rtc()?;
-
-        Ok(GolemCore::new(core))
+        Ok(core)
     }
 
     pub fn load(&mut self, program: &[u8], is_menu: bool) -> Result<GolemCore, String> {
@@ -103,20 +77,19 @@ impl CoreManager {
 
         let mut core = self.load(bytes, true)?;
 
-        if let Some(core) = core.as_any_mut().downcast_mut::<MisterFpgaCore>() {
+        if let Some(core) = core.as_any_mut().downcast_mut::<MenuCore>() {
             // Send the logo to the framebuffer.
             let logo = include_bytes!("../../../logo.png");
             let image = image::load_from_memory_with_format(logo, image::ImageFormat::Png)
                 .map_err(|e| format!("Could not load logo: {e}"))?;
 
-            let mut fullframe = image::DynamicImage::new_rgba8(1920, 1080);
+            let mut fullframe = image::DynamicImage::new_rgba8(640, 480);
             let rgba8 = fullframe.as_mut_rgba8().unwrap();
             rgba8
                 .pixels_mut()
                 .for_each(|p| *p = image::Rgba([64, 64, 64, 0]));
             image::imageops::overlay(&mut fullframe, &image, 32, 32);
-            core.menu_framebuffer_mut()?
-                .copy_from_slice(fullframe.as_bytes());
+            core.send_to_framebuffer(&fullframe.into()).unwrap();
         }
 
         self.fpga_mut().osd_enable();
