+++ src/mister-fpga/Cargo.toml
@@ -19,7 +19,7 @@ fixed-map = "0.9.3"
 glam = "0.25.0"
 hex = "0.4.3"
 i2cdev = "0.6.0"
-image = "0.25.1"
+image = "0.25"
 itertools = "0.12.0"
 json5 = "0.4.1"
 libc = "0.2.150"
