+++ src/mister-fpga/src/config/video/edid.rs
@@ -654,7 +654,7 @@ fn parse_custom_video_mode(video_mode: Option<&str>) -> CustomVideoMode {
         // return DefaultVideoMode::V1280x720r60.into();
         // return DefaultVideoMode::V640x480r60.into();
     }
-    return DefaultVideoMode::V1920x1080r60.into();
+    return DefaultVideoMode::V640x480r60.into();
 
     // todo!("parse_custom_video_mode")
 }
